refactor: consolidate text normalization utilities
